Please don't base changes on this branch
This commit introduces changes that adjust the code to work
with Tungsten Fabric paths: repository renames, Gerrit hostname,
.gitreview file etc.

As this is just a peek to the work-in-progress DO NOT PUSH real work
to this branch. This repo will likely be deleted and all work lost.

Note that `git replace` mechanism is used to show the pre-migration
history as part of the usual git log.

Signed-off-by: charlie@root.net